--- generate_examples.py
@@ -1,4 +1,0 @@
-from main import run_pipeline
-
-if __name__ == "__main__":
-    run_pipeline(["demo.txt"], "shapes.ttl", "http://example.com/atm#")
