+++ ontology_guided/__init__.py
@@ -1,0 +1,2 @@
+"""Core ontology generation modules"""
+
