+++ scripts/web_app.py
@@ -1,12 +1,15 @@
 from flask import Flask, request, render_template_string
 import os
 from werkzeug.utils import secure_filename
+
+# Import run_pipeline from the same folder
+import sys
+sys.path.append(os.path.dirname(__file__))
 from main import run_pipeline
 
 app = Flask(__name__)
 
-FORM_HTML = """
-<!doctype html>
+FORM_HTML = """<!doctype html>
 <title>Ontology Pipeline</title>
 <h1>Upload requirement file or paste text</h1>
 <form method=post enctype=multipart/form-data>
@@ -21,31 +24,35 @@ FORM_HTML = """
 {% endif %}
 """
 
-@app.route('/', methods=['GET', 'POST'])
+
+@app.route("/", methods=["GET", "POST"])
 def index():
-    if request.method == 'POST':
+    if request.method == "POST":
         inputs = []
-        os.makedirs('uploads', exist_ok=True)
-        text = request.form.get('text', '').strip()
+        os.makedirs("uploads", exist_ok=True)
+        text = request.form.get("text", "").strip()
         if text:
-            text_path = os.path.join('uploads', 'input.txt')
-            with open(text_path, 'w', encoding='utf-8') as f:
+            text_path = os.path.join("uploads", "input.txt")
+            with open(text_path, "w", encoding="utf-8") as f:
                 f.write(text)
             inputs.append(text_path)
-        uploaded = request.files.get('file')
+        uploaded = request.files.get("file")
         if uploaded and uploaded.filename:
             filename = secure_filename(uploaded.filename)
-            file_path = os.path.join('uploads', filename)
+            file_path = os.path.join("uploads", filename)
             uploaded.save(file_path)
             inputs.append(file_path)
         if not inputs:
-            return 'No input provided', 400
-        run_pipeline(inputs, 'shapes.ttl', 'http://example.com/atm#', repair=True)
-        result_path = 'results/repaired.ttl' if os.path.exists('results/repaired.ttl') else 'results/combined.ttl'
-        with open(result_path, 'r', encoding='utf-8') as f:
+            return "No input provided", 400
+        run_pipeline(inputs, "shapes.ttl", "http://example.com/atm#", repair=True)
+        result_path = (
+            "results/repaired.ttl" if os.path.exists("results/repaired.ttl") else "results/combined.ttl"
+        )
+        with open(result_path, "r", encoding="utf-8") as f:
             data = f.read()
         return render_template_string(FORM_HTML, result=data)
     return render_template_string(FORM_HTML, result=None)
 
-if __name__ == '__main__':
+
+if __name__ == "__main__":
     app.run(debug=True)
