+++ tests/test_loader.py
@@ -1,11 +1,9 @@
-import builtins
-from data_loader import DataLoader
+from ontology_guided.data_loader import DataLoader
 
 
 def test_demo_txt_loading_and_preprocessing():
     loader = DataLoader()
     texts = loader.load_requirements(["demo.txt"])
-    # Should load exactly one text from demo.txt
     assert len(texts) == 1
     sentences = []
     for t in texts:
